Add InstantMessage response mappings and largely clean up object mapping
Object mapping as a whole was rewritten quite extensively as I had come
across repositories with more information on object models.

While this is no where near complete, or even clean, its the best
working models I have been able to find due to the lack of solid
documentation on the API and object models.